(PC-25661) feat(hero): extract body to an other component
